Fix or silence (msvc) compiler warnings about unused formal parameters, unused local variables, narrowing conversions, non-copyable and/or non-instantiatible types, constant conditional expressions, and user-defined comma operators.
(cherry picked from commit 894b7ab4dcc77ec095d43ebecd236de79caec7ba)

+++ test/cuthill_mckee_ordering.cpp
@@ -63,7 +63,7 @@ int main(int , char* [])
 
   property_map<Graph,vertex_degree_t>::type deg = get(vertex_degree, G);
   for (boost::tie(ui, ui_end) = vertices(G); ui != ui_end; ++ui)
-    deg[*ui] = degree(*ui, G);
+    deg[*ui] = static_cast<int>(degree(*ui, G));
 
   property_map<Graph, vertex_index_t>::type
     index_map = get(vertex_index, G);
